code: Added "export tsv" subcommand for exporting database as TSV files
and removed TSV export functionality from "migrate" subcommand
